(broken) draft IntIntMultimap
The initial design was over-ambitious in the sense that it tries to implement two incompatible views (from key to valueset, versus a collection of key-value pairs). As a result, some attributes such as len() cannot be implemented unambiguously, leading to failed test cases.

The Markdown documentation has been updated with insights learned from failed tests. The IntIntMultimap has not yet been updated in this commit.